DOCX
Introduction:
This PowerShell script provides a simple way to track the online and offline status of devices on a network. It is designed to be easily adaptable to different network configurations and can be run as needed to update the online/offline status of devices. The script takes a list of MAC addresses and device names from an Excel file, retrieves the associated IP addresses for each MAC address, and pings each IP address to determine its online status. The results are stored in two separate arrays, one for online devices and one for offline devices. The results are then exported to a CSV file with the current date in the file name. Due to the above-mentioned nature of this approach this is a scalable and modifiable solution since adding new equipment does not require any modifications except adding the equipment to the mentioned excel document.
Prerequisites:
1.	The "ImportExcel" module must be installed on the system. (You can install it by running PowerShell as administrator and then running this command: “Install-Module -Name ImportExcel”
2.	The Excel file containing the list of MAC addresses and device names must be in the correct format. The script assumes that the columns for MAC addresses and device names are labeled "MAC-COLUMN" and "DEVICENAME-COLUMN", respectively. Replace these placeholder values with the actual column names.
3.	The system must have the necessary permissions to retrieve information about network neighbors and ping IP addresses.
Usage:
1.	Open the script in PowerShell as an administrator.
2.	Replace "FilePath" with the file path to the Excel file.
3.	Replace "YourWorkSheet" with the actual sheet name within the Excel file.
4.	Replace "MAC-COLUMN" and "DEVICENAME-COLUMN" with the actual column names for the MAC addresses and device names, respectively.
5.	Replace "SPECIFY-SAVE-LOCATION" with the desired file path for the exported CSV file.
6.	Run the script.
Output:
The script will generate two files, one for the online devices and one for the offline devices, each with the current date in the file name. The files will contain the following columns:
1.	MacAddress: The MAC address of the device.
2.	IPAddress: The IP address associated with the device. (This is omitted in the “offlinelog.csv file”)
3.	NetBIOSName: The NetBIOS name of the device.
4.	Online: Indicates whether the device is online (True) or offline (False).